Add Android "support"
Technically this doesn't add Android support, but what it does is help you through the process. Has the Exagear APK and OBB files, copies all CM files to a specific folder, and then tells you where to copy them to on your Android device. It's as close as it's going to get!