Getting Custom Database functionality working with external databases
Previously this functionality was only setup to work with databases that the user had created based on the existing ones selectable on the tool. These changes should allow a user to download any database from the Internet, and as long as it is a zip file, it should work out how to copy the files and folders properly, even if the structure isn't always the same.